+++ demisto_sdk/commands/common/errors.py
@@ -1186,8 +1186,8 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def pack_name_already_exists(new_pack_name) -> str:
-        return f"The name of your pack is: {new_pack_name} but there is an existing pack with that name, " \
-               f"please change the pack's name in the metadata file."
+        return f"A pack named: {new_pack_name} already exists in content repository, " \
+               f"change the pack's name in the metadata file."
 
     @staticmethod
     @error_code_decorator