+++ demisto_sdk/commands/common/hook_validations/pack_unique_files.py
@@ -418,7 +418,7 @@ class PackUniqueFilesValidator(BaseValidator):
 
         return True
 
-    def are_valid_files(self) -> str:
+    def are_valid_files(self, id_set_validations) -> str:
         """Main Execution Method"""
         self.validate_secrets_file()
         self.validate_pack_ignore_file()
@@ -428,6 +428,13 @@ class PackUniqueFilesValidator(BaseValidator):
         # We only check pack dependencies for -g flag
         if self.validate_dependencies:
             self.validate_pack_dependencies()
+
+        # Check if unique files are valid against the rest of the files, using the ID set.
+        if id_set_validations:
+            is_valid, error = id_set_validations.is_unique_file_valid_in_set(self.pack_path, self.ignored_errors)
+            if not is_valid:
+                self._add_error(error, self.pack_path)
+
         return self.get_errors()
 
     # pack dependencies validation