+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 * Added the **doc-review** command to check spelling in .md and .yml files as well as a basic release notes review.
-* Added a validation that pack display name did not exist in our repo.
+* Added a validation that a pack's display name does not already exist in content repository.
 
 # 1.3.1
 * Fixed an issue where the **validate** command failed to validate the release notes of beta integrations.