+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Updated the format of the outputs when using the *--json-file* flag to create a JSON file output for the **validate** and **lint** commands.
 * Added the **doc-review** command to check spelling in .md and .yml files as well as a basic release notes review.
 
 # 1.3.1